+++ examples/menu.jl
@@ -19,6 +19,9 @@ using REPL.TerminalMenus
 
 options = [
     "v3kite = include(\"v3kite.jl\")",
+    "v3beam = include(\"v3beam.jl\")",
+    "v3beam_geometry = include(\"v3beam_geometry.jl\")",
+    "relax_bridle = include(\"relax_bridle.jl\")",
     "reel_out_v3 = include(\"reel_out_v3.jl\")",
     "open_loop = include(\"open_loop.jl\")",
     "realtime = include(\"realtime.jl\")",